Add schedule button to posts list
