+++ app/components/user-management.tsx
@@ -39,32 +39,7 @@ type User = {
 };
 
 export default function UserManagement() {
-  const [users, setUsers] = useState<Array<User>>([
-    // {
-    //   id: 1,
-    //   name: "John Doe",
-    //   email: "john@example.com",
-    //   role: "user",
-    //   plan: "pro",
-    //   activity: "active",
-    // },
-    // {
-    //   id: 2,
-    //   name: "Jane Smith",
-    //   email: "jane@example.com",
-    //   role: "admin",
-    //   plan: "enterprise",
-    //   activity: "inactive",
-    // },
-    // {
-    //   id: 3,
-    //   name: "Bob Johnson",
-    //   email: "bob@example.com",
-    //   role: "user",
-    //   plan: "basic",
-    //   activity: "active",
-    // },
-  ]);
+  const [users, setUsers] = useState<Array<User>>([]);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
   const [searchTerm, setSearchTerm] = useState("");
